# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Deesha Foundation API
  version: 1.0.0
  description: |
    API for Deesha Foundation - donation management, event registration, and payment processing.
    
    ## Authentication
    Most endpoints require authentication via Supabase Auth JWT tokens.
    
    ## Rate Limiting
    Public endpoints are rate-limited. See individual endpoint documentation for limits.
  contact:
    name: Deesha Foundation
    url: https://deesha.org
  license:
    name: Proprietary
    url: https://deesha.org
servers:
  - url: https://deesha.org/api
    description: Production
  - url: https://staging.deesha.org/api
    description: Staging
  - url: http://localhost:3000/api
    description: Development
...
paths:
...
  /events/start-payment:
    post:
      tags: [Events]
      summary: Start event payment
      description: Creates a payment session for event registration
      operationId: startEventPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [registrationId, email, provider]
              properties:
                registrationId:
                  type: string
                  format: uuid
                email:
                  type: string
                  format: email
                provider:
                  type: string
                  enum: [stripe, khalti, esewa]
      responses:
        '200':
          description: Payment session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      paymentUrl:
                        type: string
                        format: uri
                      provider:
                        type: string
                      sessionId:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'